Use same dependencies as GitHub Pages runs with
By using the GitHub Pages gem you transitively get the versions of
dependencies that GHP will use in production. This reduces surprises
when deploying.

.ruby-version will be read by Ruby version managers (such as chruby,
rvm, rbenv) so that the version of Ruby used is at best more consistent
and at worst version controlled. It won’t change the version that GHP
uses, but it will version control the local environment for developers
who use a version manager.