Fix warnings due to liquid syntax errors
After using the same dependencies as GHP, the following errors were
shown. These are due to both incorrect use of the default filter (the
default value should follow a colon rather than being wrapped in
brackets), and incorrect checking for definition of variables. In
liquid, variables are ‘truthy’ when they are defined, so the use of
defaults isn’t necessary when using them in a conditional in this way.